Clearly defined interface between front and back.
What
===
Redefined the interface between the front and backend so it was
intuitive and didnt use idioms from Imgur.

Why
===
The backend will make calls out to multiple APIs, not just Imgur, and
it's better if the frontend knows nothing about Imgur to start with
before adding those other integrations.