Moved semi-private information to env.
What
===
Moved semi-private information to env.

Why
===
I accidentally committed the Imgur Client ID in an earlier commit when
this project was mostly a hack for fun. It's not really private, it's
not the secret key, but it's not ideal and should be in an environment
variable in any case.