Page load with query in URL searches immediately.
What
===
Page load with query in URL searches immediately.

Why
===
It would delay in the past, but this was only necessary before because
we needed to make sure there was a connection to the web service first.
Now the web service has been removed we can hit the giphy API
immediately from the browser.